Nft api call refactored in VM & views
