+++ CryptoNoob/View/FavoritesView/FavoritesSheetView.swift
@@ -35,7 +35,6 @@ struct FavoritesSheetView: View {
 struct FavoritesSheetView_Previews: PreviewProvider {
     static var previews: some View {
         FavoritesSheetView()
-			.environmentObject(ApiCall())
 			.environmentObject(FavoriteViewModel())
     }
 }
