+++ CryptoNoob/View/MainView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct MainView: View {
-    @EnvironmentObject var apiCall: ApiCall
+	@EnvironmentObject var fetchNft: FetchNftApi
     @State private var isOn: Bool = false
     @State private var selection = 0
 
@@ -33,7 +33,7 @@ struct MainView_Previews: PreviewProvider {
     static var previews: some View {
         MainView()
             .preferredColorScheme(.dark)
-            .environmentObject(ApiCall())
+            .environmentObject(FetchNftApi())
     }
 }
